post storage

+++ app/Http/Requests/StorePostRequest.php
@@ -26,7 +26,7 @@ class StorePostRequest extends FormRequest
         return [
             'title' => ['string','required','max:60'],
             'contenido' => ['string','required','max:255'],
-            'community'=>['int'],
+            'community'=>['numeric'],
         ];
     }
 }
